Revert Merge branch 'master' of git://github.com/files-community/files-uwp into localization
