+++ Files.Launcher/Files.Launcher.csproj
@@ -134,7 +134,7 @@
       <Version>4.7.0</Version>
     </PackageReference>
     <PackageReference Include="Vanara.Windows.Shell">
-      <Version>3.2.15</Version>
+      <Version>3.2.16</Version>
     </PackageReference>
   </ItemGroup>
   <ItemGroup>
