--- Files/Dialogs/ExtractFilesDialog.xaml
@@ -1,42 +1,0 @@
-﻿<ContentDialog
-    x:Class="Files.Dialogs.ExtractFilesDialog"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local2="using:Files.Helpers"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    x:Uid="ExtractFilesDialog"
-    Title="Extract Compressed Archive"
-    CloseButtonClick="ContentDialog_CloseButtonClick"
-    CloseButtonText="Cancel"
-    CornerRadius="4"
-    DefaultButton="Primary"
-    PrimaryButtonClick="ContentDialog_PrimaryButtonClick"
-    PrimaryButtonText="Extract"
-    RequestedTheme="{x:Bind local2:ThemeHelper.RootTheme}"
-    mc:Ignorable="d">
-
-    <Grid MinWidth="375">
-        <StackPanel Orientation="Vertical">
-            <TextBlock
-                x:Uid="ExtractFilesDialogDescription"
-                Text="Pick a location to extract this compressed archive to. You'll need to stay in the current folder until we're done. A new tab will open up with the extracted items."
-                TextWrapping="WrapWholeWords" />
-            <Grid Margin="0,14,0,0" ColumnSpacing="5">
-                <Grid.ColumnDefinitions>
-                    <ColumnDefinition Width="75*" />
-                    <ColumnDefinition Width="25*" />
-                </Grid.ColumnDefinitions>
-                <TextBox x:Name="DestPathText" Grid.Column="0" />
-                <Button
-                    x:Name="BrowseButton"
-                    x:Uid="ExtractFilesDialogBrowseButton"
-                    Grid.Column="1"
-                    HorizontalAlignment="Stretch"
-                    VerticalAlignment="Stretch"
-                    Click="BrowseButton_Click"
-                    Content="Browse" />
-            </Grid>
-        </StackPanel>
-    </Grid>
-</ContentDialog>
