+++ Files/Interacts/Interaction.cs
@@ -1493,88 +1493,6 @@ namespace Files.Interacts
             AddItemDialog.CreateFile(AddItemType.BitmapImage);
         }
 
-        public async void ExtractItems_Click(object sender, RoutedEventArgs e)
-        {
-            var selectedIndex = CurrentInstance.ContentPage.GetSelectedIndex();
-            StorageFile selectedItem = await ItemViewModel.GetFileFromPathAsync(CurrentInstance.FilesystemViewModel.FilesAndFolders[selectedIndex].ItemPath);
-
-            StatusBanner banner = null;
-            ExtractFilesDialog extractFilesDialog = new ExtractFilesDialog(CurrentInstance.FilesystemViewModel.WorkingDirectory);
-            await extractFilesDialog.ShowAsync();
-            if (((bool)ApplicationData.Current.LocalSettings.Values["Extract_Destination_Cancelled"]) == false)
-            {
-                var bufferItem = await selectedItem.CopyAsync(ApplicationData.Current.TemporaryFolder, selectedItem.DisplayName, NameCollisionOption.ReplaceExisting);
-                string destinationPath = ApplicationData.Current.LocalSettings.Values["Extract_Destination_Path"].ToString();
-                StorageFolder destinationFolder = await ItemViewModel.GetFolderFromPathAsync(destinationPath);
-                //ZipFile.ExtractToDirectory(selectedItem.Path, destinationPath, );
-                var destFolder_InBuffer = await ApplicationData.Current.TemporaryFolder.CreateFolderAsync(selectedItem.DisplayName + "_Extracted", CreationCollisionOption.ReplaceExisting);
-                using FileStream fs = new FileStream(bufferItem.Path, FileMode.Open);
-                ZipArchive zipArchive = new ZipArchive(fs);
-                int totalCount = zipArchive.Entries.Count;
-                int index = 0;
-
-                App.InteractionViewModel.IsContentLoadingIndicatorVisible = false;
-                banner = App.CurrentInstance.StatusBarControl.OngoingTasksControl.PostBanner(
-                    null,
-                    App.CurrentInstance.FilesystemViewModel.WorkingDirectory,
-                    0,
-                    StatusBanner.StatusBannerSeverity.Ongoing,
-                    StatusBanner.StatusBannerOperation.Extract);
-
-                await Task.Run((Func<Task>)(async () =>
-                {
-                    foreach (ZipArchiveEntry archiveEntry in zipArchive.Entries)
-                    {
-                        uint progressValue = (uint)(index * 100.0 / zipArchive.Entries.Count);
-                        await CoreApplication.MainView.CoreWindow.Dispatcher.RunAsync(CoreDispatcherPriority.Normal,
-                            new DispatchedHandler(() =>
-                            {
-                                if (banner == null)
-                                {
-                                    throw new ArgumentNullException();
-                                }
-                                else
-                                {
-                                    banner.Report((uint)progressValue);
-                                }
-                            }));
-                        if (archiveEntry.FullName.Contains('/'))
-                        {
-                            var nestedDirectories = archiveEntry.FullName.Split('/').ToList();
-                            nestedDirectories.Remove(nestedDirectories.Last());
-                            var relativeOutputPathToEntry = Path.Combine(nestedDirectories.ToArray());
-                            System.IO.Directory.CreateDirectory(Path.Combine(destFolder_InBuffer.Path, relativeOutputPathToEntry));
-                        }
-
-                        if (!string.IsNullOrWhiteSpace(archiveEntry.Name))
-                            archiveEntry.ExtractToFile(Path.Combine(destFolder_InBuffer.Path, archiveEntry.FullName));
-
-                        index++;
-                        if (index == totalCount)
-                        {
-                            App.InteractionViewModel.IsContentLoadingIndicatorVisible = false;
-                        }
-                    }
-                }));
-
-                await CloneDirectoryAsync(destFolder_InBuffer, destinationFolder, destFolder_InBuffer.Name, true)
-                    .ContinueWith(async (x) =>
-                {
-                    await destFolder_InBuffer.DeleteAsync(StorageDeleteOption.PermanentDelete);
-                    await CoreApplication.MainView.CoreWindow.Dispatcher.RunAsync(CoreDispatcherPriority.Normal, async () =>
-                    {
-                        await MainPage.AddNewTab(typeof(ModernShellPage), destinationPath + "\\" + selectedItem.DisplayName + "_Extracted");
-                    });
-                });
-                banner.Report(100);
-                App.CurrentInstance.StatusBarControl.OngoingTasksControl.RemoveBanner(banner);
-            }
-            else if (((bool)ApplicationData.Current.LocalSettings.Values["Extract_Destination_Cancelled"]) == true)
-            {
-                return;
-            }
-        }
-
         public void SelectAllItems() => CurrentInstance.ContentPage.SelectAllItems();
 
         public void InvertAllItems() => CurrentInstance.ContentPage.InvertSelection();
