+++ Files/Strings/fr-FR/Resources.resw
@@ -171,15 +171,6 @@
   <data name="NavigationToolbarNewTab.Text" xml:space="preserve">
     <value>Nouvel onglet</value>
   </data>
-  <data name="ExtractFilesDialog.CloseButtonText" xml:space="preserve">
-    <value>Annuler</value>
-  </data>
-  <data name="ExtractFilesDialog.PrimaryButtonText" xml:space="preserve">
-    <value>Extraire</value>
-  </data>
-  <data name="ExtractFilesDialogBrowseButton.Content" xml:space="preserve">
-    <value>W!!ßrøwsè 表!!W</value>
-  </data>
   <data name="BaseLayoutContextFlyoutSortBy.Text" xml:space="preserve">
     <value>Trier la liste par</value>
   </data>
@@ -533,9 +524,6 @@
   </data>
   <data name="DeleteInProgress.Title" xml:space="preserve">
     <value>Suppression des fichiers</value>
-  </data>
-  <data name="ExtractInProgress.Title" xml:space="preserve">
-    <value>Extraction de fichiers...</value>
   </data>
   <data name="BitlockerDialog.PrimaryButtonText" xml:space="preserve">
     <value>Déverrouiller</value>
