+++ Files/Strings/tr-TR/Resources.resw
@@ -174,21 +174,6 @@
   <data name="NavigationToolbarNewTab.Text" xml:space="preserve">
     <value>Yeni Sekme</value>
   </data>
-  <data name="ExtractFilesDialog.Title" xml:space="preserve">
-    <value>Sıkıştırılmış Arşivi Çıkart</value>
-  </data>
-  <data name="ExtractFilesDialog.CloseButtonText" xml:space="preserve">
-    <value>İptal</value>
-  </data>
-  <data name="ExtractFilesDialog.PrimaryButtonText" xml:space="preserve">
-    <value>Çıkart</value>
-  </data>
-  <data name="ExtractFilesDialogDescription.Text" xml:space="preserve">
-    <value>Bu sıkıştırılmış arşivi çıkarmak için bir konum seçin. İşlem tamamlanana kadar geçerli klasörde kalmanız gerekir. Çıkarılan öğelerle yeni bir sekme açılacaktır.</value>
-  </data>
-  <data name="ExtractFilesDialogBrowseButton.Content" xml:space="preserve">
-    <value>Gözat</value>
-  </data>
   <data name="BaseLayoutContextFlyoutSortBy.Text" xml:space="preserve">
     <value>Göre Sırala</value>
   </data>
@@ -572,8 +557,5 @@
   </data>
   <data name="DeleteInProgress.Title" xml:space="preserve">
     <value>Dosyalar siliniyor</value>
-  </data>
-  <data name="ExtractInProgress.Title" xml:space="preserve">
-    <value>Dosyalar ayıklanıyor ...</value>
   </data>
 </root>
