+++ Files/Strings/zh-Hans/Resources.resw
@@ -174,21 +174,6 @@
   <data name="NavigationToolbarNewTab.Text" xml:space="preserve">
     <value>新建标签页</value>
   </data>
-  <data name="ExtractFilesDialog.Title" xml:space="preserve">
-    <value>提取压缩文档</value>
-  </data>
-  <data name="ExtractFilesDialog.CloseButtonText" xml:space="preserve">
-    <value>取消</value>
-  </data>
-  <data name="ExtractFilesDialog.PrimaryButtonText" xml:space="preserve">
-    <value>提取</value>
-  </data>
-  <data name="ExtractFilesDialogDescription.Text" xml:space="preserve">
-    <value>选择一个位置以提取此压缩存档。提取完成前请不要离开此页，完成后将打开新标签页显示提取的文件。</value>
-  </data>
-  <data name="ExtractFilesDialogBrowseButton.Content" xml:space="preserve">
-    <value>浏览</value>
-  </data>
   <data name="BaseLayoutContextFlyoutSortBy.Text" xml:space="preserve">
     <value>排序方式</value>
   </data>
@@ -656,9 +641,6 @@
   </data>
   <data name="DeleteInProgress.Title" xml:space="preserve">
     <value>删除文件</value>
-  </data>
-  <data name="ExtractInProgress.Title" xml:space="preserve">
-    <value>提取文件</value>
   </data>
   <data name="PasteInProgress.Title" xml:space="preserve">
     <value>粘贴文件</value>
