+++ Files/UserControls/FileIcon.xaml
@@ -1,0 +1,67 @@
+﻿<UserControl
+    x:Class="Files.UserControls.FileIcon"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Files.UserControls"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d"
+    d:DesignHeight="300"
+    d:DesignWidth="400">
+
+    <StackPanel>
+        <Image
+            x:Name="FolderGlyphIcon"
+            Width="{x:Bind ItemSize}"
+            Height="{x:Bind ItemSize}"
+            HorizontalAlignment="Stretch"
+            VerticalAlignment="Stretch"
+            x:Load="{x:Bind ViewModel.LoadFolderGlyph}"
+            Stretch="Uniform">
+            <Image.Source>
+                <SvgImageSource
+                    RasterizePixelHeight="128"
+                    RasterizePixelWidth="128"
+                    UriSource="{x:Bind ViewModel.FolderIconSource}" />
+            </Image.Source>
+        </Image>
+        <FontIcon
+            x:Name="EmptyImageIcon"
+            HorizontalAlignment="Left"
+            VerticalAlignment="Stretch"
+            x:Load="{x:Bind ViewModel.LoadUnknownTypeGlyph, Mode=OneWay}"
+            FontFamily="{StaticResource FluentUIGlyphs}"
+            FontSize="{x:Bind ItemSize}"
+            Glyph="&#xea00;" />
+        <FontIcon
+            x:Name="CombinedItemsIcon"
+            HorizontalAlignment="Left"
+            VerticalAlignment="Stretch"
+            x:Load="{x:Bind ViewModel.LoadCombinedItemsGlyph, Mode=OneWay}"
+            FontFamily="{StaticResource CustomGlyph}"
+            FontSize="{x:Bind LargerItemSize}"
+            Glyph="&#xF109;" />
+        <FontIcon
+            x:Name="DriveItemIcon"
+            HorizontalAlignment="Center"
+            VerticalAlignment="Stretch"
+            x:Load="{x:Bind ViewModel.LoadDriveItemGlyph, Mode=OneWay}"
+            FontFamily="{StaticResource FluentUIGlyphs}"
+            FontSize="{x:Bind LargerItemSize}"
+            Glyph="{x:Bind ViewModel.DriveItemGlyphSource, Mode=OneWay}" />
+        <FontIcon
+            x:Name="WebShortcutGlyph"
+            HorizontalAlignment="Left"
+            VerticalAlignment="Stretch"
+            x:Load="{x:Bind ViewModel.LoadLinkIcon, Mode=OneWay}"
+            FontFamily="{StaticResource FluentUIGlyphs}"
+            FontSize="{x:Bind ItemSize}"
+            Glyph="&#xEAAB;" />
+        <Image
+            x:Name="ItemIcon"
+            HorizontalAlignment="Stretch"
+            VerticalAlignment="Stretch"
+            x:Load="{x:Bind ViewModel.LoadFileIcon, Mode=OneWay}"
+            Source="{x:Bind ViewModel.FileIconSource, Mode=OneWay}" />
+    </StackPanel>
+</UserControl>
