+++ Files/Views/Pages/PropertiesGeneral.xaml.cs
@@ -1,68 +1,17 @@
 using Files.Filesystem;
-using Files.View_Models;
 using Files.View_Models.Properties;
 using Microsoft.Toolkit.Uwp.Helpers;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 using Windows.ApplicationModel.Core;
-using Windows.Storage;
-using Windows.UI.Xaml;
-using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Navigation;
-
-// Il modello di elemento Pagina vuota è documentato all'indirizzo https://go.microsoft.com/fwlink/?LinkId=234238
 
 namespace Files
 {
-    /// <summary>
-    /// Pagina vuota che può essere usata autonomamente oppure per l'esplorazione all'interno di un frame.
-    /// </summary>
-    public sealed partial class PropertiesGeneral : Page
+    public sealed partial class PropertiesGeneral : PropertiesTab
     {
-        public BaseProperties BaseProperties { get; set; }
-
-        public SelectedItemsPropertiesViewModel ViewModel { get; set; }
-
         public PropertiesGeneral()
         {
             this.InitializeComponent();
-        }
-
-        private void Properties_Loaded(object sender, RoutedEventArgs e)
-        {
-            if (BaseProperties != null)
-            {
-                BaseProperties.GetSpecialProperties();
-            }
-        }
-
-        protected override void OnNavigatedTo(NavigationEventArgs e)
-        {
-            ViewModel = new SelectedItemsPropertiesViewModel();
-            var np = e.Parameter as Properties.PropertyNavParam;
-
-            if (np.navParameter is ListedItem)
-            {
-                var listedItem = np.navParameter as ListedItem;
-                if (listedItem.PrimaryItemAttribute == StorageItemTypes.File)
-                {
-                    BaseProperties = new FileProperties(ViewModel, np.tokenSource, Dispatcher, ItemMD5HashProgress, listedItem);
-                }
-                else if (listedItem.PrimaryItemAttribute == StorageItemTypes.Folder)
-                {
-                    BaseProperties = new FolderProperties(ViewModel, np.tokenSource, Dispatcher, listedItem);
-                }
-            }
-            else if (np.navParameter is List<ListedItem>)
-            {
-                BaseProperties = new CombinedProperties(ViewModel, np.tokenSource, Dispatcher, np.navParameter as List<ListedItem>);
-            }
-            else if (np.navParameter is DriveItem)
-            {
-                BaseProperties = new DriveProperties(ViewModel, np.navParameter as DriveItem);
-            }
-
-            base.OnNavigatedTo(e);
+            base.ItemMD5HashProgress = ItemMD5HashProgress;
         }
 
         public async Task SaveChanges(ListedItem item)
