+++ README.md
@@ -1,7 +1,8 @@
 ## Files UWP
 Meet Files, an enthusiast take on what Windows File Explorer explorer <b>should</b> be.
 <br/><br/>
-<a href="https://www.microsoft.com/store/apps/9NGHP3DX8HDX">Download Files UWP from the Microsoft Store.</a>
+<a href="https://www.microsoft.com/store/apps/9NGHP3DX8HDX">Download Files UWP</a>
+<p>Files UWP </p><a href="https://github.com/files-community/files-uwp/blob/master/CODE_OF_CONDUCT.md">code of conduct</a>
 
 [![Build Status](https://dev.azure.com/lukeblevins150823/Files%20UWP/_apis/build/status/Build%20Pipeline?branchName=master)](https://dev.azure.com/lukeblevins150823/Files%20UWP/_build/latest?definitionId=4&branchName=master)
 [![Discord](https://discordapp.com/api/guilds/725513575971684472/widget.png)](https://discord.gg/mr5hVu8)
